pyproject.toml
[project]
name = "customer-ontology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.25",
    "lxml>=5.0.0",
    "openai>=1.10.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "xlsxwriter>=3.1.0",
]
